% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_test_summary.r
\name{adf_test_summary}
\alias{adf_test_summary}
\title{Run ADF test for unit roots on all columns of a data frame.}
\usage{
adf_test_summary(df, p_select = "BIC")
}
\arguments{
\item{p_select}{Method for selecting number of lags to include in the test regression. Possible values: "AIC", "BIC"}
}
\value{
A data frame with 3 rows and ncol(df) columns.
}
\description{
Run ADF test for unit roots on all columns of a data frame.
}
\examples{
# Example 1: We test it on iid white noises and a random walk.
# As expected, white_noise's p-value is smaller than 0.01, so it is stationary. But random walk is not.
set.seed(10)
df <- data.frame(white_noise = rnorm(100), random_walk = cumsum(rnorm(100)))
adf_test_summary(df)

# Example 2: We test it on a "trend stationary" time series.
set.seed(10)
df <- data.frame(trend_s = 0.8 * seq(1:20) + rnorm(20))
adf_test_summary(df)
}
