% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_rate.R
\name{missing_rate}
\alias{missing_rate}
\title{Summarize the missing rate of a data frame}
\usage{
missing_rate(df, include_non_missing = FALSE)
}
\description{
This function summarizes the missing rate of the given data frame. It simply computes the number of rows having NA (as
well as the percentage) under each column. The result is sorted by percentage of missing values.
}
